Templating improvements
Standardized a few minor layout discrepencies across pages. The template now wraps content (i.e. anything that's not the navigation bar or the footer) in the 'main' div, which has some very basic styling to make it play nice.

Sign-in option only appears if user is not signed in.

The notification page now displays without any errors. Cleaned up the templating which was crashing the page.