User menu
Added major navigation functionality by implementing a user dropdown menu, such that a logged in user can navigate directly to their profile or notifications, as well as being able to log out or add a book.

This makes navigation a lot easier and improves the ui by making navigation cues more explicit. It also fills gaps such as navigating to notifications and logging out, which a user cannot easily do at the moment.

Logged out users will no longer see the profile link, only "log in" and "sign up" buttons.

A possible extension is to add an indicator to the icon if the user has notifications. Lmk if I should pursue this.