Update DockerFile
Update the DockerFile to include the production environment and using ARG to separate production and testing cycle. 
As we do not have a separate testing environment, I will leave it as a comment in the file. But when we scale up and have testing environment variables (test.env), we can just add into this DockerFile.

Link for future reference: https://dev.to/stevoperisic/dynamic-dockerfile-with-arg-2h62